(feat) export repetitive processes to utils and refactor files

+++ server/routes/address.js
@@ -1,26 +1,24 @@
-const router = require("express").Router();
+const router = require('express').Router();
+const axios = require('axios');
+const XLSX = require('xlsx');
+const rimraf = require('rimraf');
+const multer = require('multer');
+const {
+  apiLinkGenerator,
+  queryResult,
+  errorWithResponse,
+  errorWithoutResponse
+} = require('../utils');
 module.exports = router;
-const axios = require("axios");
-if (process.env.NODE_ENV !== "production") require("../../secrets");
-const apiKey = process.env.API_KEY;
-const XLSX = require("xlsx");
-const multer = require("multer");
-const rimraf = require("rimraf");
 
-//saves file locally for processing
 const storage = multer.diskStorage({
   destination: "./uploads",
   filename(req, file, cb) {
-    cb(null, `${new Date()}-${file.originalname}`);
+    cb(null, `${new Date().getTime()}-${file.originalname}`);
   }
 });
 
 const upload = multer({ storage }).single("file");
-
-//formats api call
-const addresslinkGenerator = (key, address) => {
-  return `https://api.trade.gov/consolidated_screening_list/search?api_key=${key}&address=${address}`;
-};
 
 let spreadsheetForAnalysis, data, worksheet;
 let apiInput = [];
@@ -33,62 +31,55 @@ router.post("/spreadsheet", upload, (req, res, next) => {
     spreadsheetForAnalysis = XLSX.readFile(file.path);
     worksheet = spreadsheetForAnalysis.Sheets[spreadsheetForAnalysis.SheetNames[0]];
     data = XLSX.utils.sheet_to_json(worksheet, { header: 1 }, { raw: false });
-    data = data.filter(
-      cellContent => cellContent[0] && cellContent[0].length > 1);
+    data = data.filter(cellContent => cellContent[0] && cellContent[0].length > 1);
+    //this removes the column label row
     data.shift();
     res.send({ listlength: data.length })
   } else {
-      res.send('Only excel file inputs allowed'),
-      rimraf('uploads', () => console.log(' uploads removed'))
+    res.send('Only excel file inputs allowed'),
+    rimraf('uploads', () => console.log(' uploads removed'))
   }
-    next()
-  });
+  next()
+});
 
 router.delete("/spreadsheet/reset", (req, res, next) => {
-  rimraf('uploads', () => console.log(' uploads removed'))  
   finalAddressResults = [];
   res.send({ listlength: 0 });
+  rimraf('uploads', (err) => console.log('ready for new list'));
+  next()
 });
 
 router.post("/", (req, res, next) => {
   let i = Number(req.body.count);
   let j = i + 100;
   apiInput = data.slice(i, j);
+  //remove temp upload directory since file is now processed under 'data'
+  if (i < 100) rimraf('uploads', () => console.log(' uploads removed'))
+
   while (i < data.length) {
     return Promise.all(
       apiInput.map(query => {
         const address = query[0];
         return axios
-          .get(addresslinkGenerator(apiKey, query[0]))
+          .get(apiLinkGenerator(`address`, address))
           .then(result => {
-            const formattedReturn = {
-              addressSearched: address,
-              data: result.data,
-              api: addresslinkGenerator(apiKey, query[0])
-            };
-            finalAddressResults.push(formattedReturn);
+            finalAddressResults.push(queryResult(`address`, address, result));
           })
           .catch(err =>
-            finalAddressResults.push({
-              addressSearched: address,
-              error: {
-                message: err.response
-                  ? `${err.response.status}, ${err.response.satusText}`
-                  : `error response malformed`,
-                url: addresslinkGenerator(apiKey, query[0])
-              }
-            })
+            err.response ?
+              finalAddressResults.push(errorWithResponse(address, err))
+              :
+              finalAddressResults.push(errorWithoutResponse(`address`, address))
           );
       })
     )
-      .then(() => (
-        (j = i + 100), 
-        res.send(finalAddressResults),
-        rimraf('uploads', () => console.log(' uploads removed'))  
-        ))
-      .catch(err => {
-        res.status(500).send(err);
-        console.log("=============== fetch data error", err);
-      });
+    .then(() => (
+      (j = i + 100),
+      res.send(finalAddressResults)
+    ))
+    .catch(err => {
+      res.status(500).send(err);
+      console.log("=============== fetch data error", err);
+    });
   }
 });
